DOMA-3952 review fixes - move constants, extract cookie parsing to function & use report type constant at test utils
