+++ apps/condo/domains/analytics/components/TicketChartView.tsx
@@ -13,7 +13,7 @@ import { CHART_COLOR_SET } from '@condo/domains/common/constants/style'
 import { TicketGroupedCounter } from '@app/condo/schema'
 import { colors } from '@condo/domains/common/constants/style'
 import InfiniteScroll from 'react-infinite-scroller'
-import { TICKET_CHART_PAGE_SIZE } from '@condo/domains/analytics/constants/restrictions'
+import { TICKET_CHART_PAGE_SIZE } from '@condo/domains/analytics/utils/helpers'
 import { getChartOptions, GroupTicketsByTypes } from '@condo/domains/analytics/utils/helpers'
 import styled from '@emotion/styled'
 import { Button } from '@condo/domains/common/components/Button'
