+++ apps/condo/domains/ticket/constants/restrictions.ts
@@ -1,8 +1,5 @@
 const MAX_DESCRIPTION_DISPLAY_LENGTH = 150
 const MIN_DESCRIPTION_LENGTH = 6
-const TICKET_CHART_PAGE_SIZE = 6
-const MAX_CHART_NAME_LENGTH = 35
-const MAX_CHART_LEGEND_ELEMENTS = 2
 const MAX_FILTERED_ELEMENTS = 20
 const MAX_TAG_TEXT_LENGTH = 30
 const MAX_PHONE_LENGTH = 15
@@ -11,9 +8,6 @@ const MIN_PHONE_LENGTH = 11
 export {
     MAX_DESCRIPTION_DISPLAY_LENGTH,
     MIN_DESCRIPTION_LENGTH,
-    TICKET_CHART_PAGE_SIZE,
-    MAX_CHART_NAME_LENGTH,
-    MAX_CHART_LEGEND_ELEMENTS,
     MAX_FILTERED_ELEMENTS,
     MAX_TAG_TEXT_LENGTH,
     MAX_PHONE_LENGTH,
